Use watgpu308 instead
The trick to schedule jobs on the RTX 6000 Ada when there are other GPUs on the machine is to temporarily reserve all the GPUs before it, then allocate the job, then quit the initial reservation. The scheduler always seems to pick the first available GPU by PCIe ordering (same ordering as in nvidia-smi).